Comenzamos a modificar la forma en que se crean lso nuevos archivos desde los xml

+++ src/helpers/files-helpers.ts
@@ -19,9 +19,19 @@ export const createFile = (data: {
 	name?: string;
 	extension?: FileType;
 }): string => {
-	const { extension, name, content, path = 'src/examples/' } = data;
+	const { extension, name, content, path = '_generated' } = data;
 	const fileName = `${name}.${extension}`;
 	const fullPath = `${path}/${fileName}`;
+
+	if (!fs.existsSync(path)) {
+		fs.mkdirSync(path, { recursive: true });
+		console.log(`Directory '${path}' created`);
+	  }
+
+	if (fs.existsSync(fullPath)) {
+		console.error(`Error: File '${fullPath}' already exists`);
+		return fullPath;
+	}
 	fs.writeFileSync(fullPath, content);
-	return `File created at ${fullPath}`;
+	return fullPath;
 };
